% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preform_operations.r
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Add two numbers together}
\usage{
add_two_numbers(first_number, second_number)
}
\arguments{
\item{first_number}{first number to be added}

\item{second_number}{second number to be added}
}
\value{
addition of two inputs
}
\description{
Take two numbers as inputs as add them
}
