% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preform_operations.r
\name{arithmetic_with_multiple_numbers}
\alias{arithmetic_with_multiple_numbers}
\title{perform operation on multiple numbers}
\usage{
arithmetic_with_multiple_numbers(
  operator = c("add", "subtract", "multiply", "divide"),
  first_number,
  ...
)
}
\arguments{
\item{operator}{: {"add", "subtract", "multiply", "divide"} Operation to do with `first_number` and `second_number`}

\item{first_number}{Number to add, subtract, multiply or divide}

\item{...}{allows for any number of numbers to be added as input variables.}
}
\value{
output from operation
}
\description{
The sum, difference, product or quotient of multiple numbers.
}
