% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preform_operations.r
\name{arithmetic_with_two_numbers}
\alias{arithmetic_with_two_numbers}
\title{Perform operation on two numbers}
\usage{
arithmetic_with_two_numbers(
  first_number,
  second_number,
  operator = c("add", "subtract", "multiply", "divide")
)

arithmetic_with_two_numbers(
  first_number,
  second_number,
  operator = c("add", "subtract", "multiply", "divide")
)
}
\arguments{
\item{first_number}{Number to add, subtract, multiply or divide}

\item{second_number}{Number to add, subtract, multiply or divide from first}

\item{operator}{: {"add", "subtract", "multiply", "divide"} Operation to do with `first_number` and `second_number`}
}
\value{
output from operation

output from operation
}
\description{
The sum, difference, product or quotient of two numbers.

The sum, difference, product or quotient of two numbers.
}
