% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preform_operations.r
\name{divide_two_numbers}
\alias{divide_two_numbers}
\title{divide two numbers}
\usage{
divide_two_numbers(first_number, second_number)
}
\arguments{
\item{first_number}{first number}

\item{second_number}{second number to be divided from first}
}
\value{
addition of two inputs
}
\description{
Take two numbers as inputs and divide them
}
