% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobbyist.R
\name{hobbyist_fun}
\alias{hobbyist_fun}
\title{Making it a fun}
\usage{
hobbyist_fun()
}
\description{
Lah-de-dah! Say no More!!!  
Here we create a function that makes use of rio::import to read a csv datafile 
and ggplot2 to render graphics of dataset loaded.
}
\examples{
 
#Not Run  
hobbyist_fun()  
?hobbyist_fun
}
