% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preform_operations.r
\name{multiply_two_numbers}
\alias{multiply_two_numbers}
\title{multiply two numbers together}
\usage{
multiply_two_numbers(first_number, second_number)
}
\arguments{
\item{first_number}{first number to be multiplied}

\item{second_number}{second number to be multiplied}
}
\value{
multiplication of two inputs
}
\description{
Take two numbers as inputs and multiply them
}
