% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preform_operations.r
\name{perform_operation_api}
\alias{perform_operation_api}
\title{perform multiple operations}
\usage{
perform_operation_api(first_number, second_number)
}
\arguments{
\item{first_number}{first number to be added}

\item{second_number}{second number to be added}
}
\value{
list of operation outputs
}
\description{
Perform adding, subtrating, multiplying and dividing on numbers as an API
}
