% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_config}
\alias{read_config}
\title{read in config}
\usage{
read_config(directory = "config.yaml")
}
\arguments{
\item{directory}{directory set to top level where config sits}
}
\value{
Config
}
\description{
Read in config from top level
}
