% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preform_operations.r
\name{subtract_two_numbers}
\alias{subtract_two_numbers}
\title{subtract two numbers}
\usage{
subtract_two_numbers(first_number, second_number)
}
\arguments{
\item{first_number}{first number}

\item{second_number}{number to subtract from first number}
}
\value{
subtraction of inputs
}
\description{
Take two numbers as inputs and subtract them
}
