% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy-function.R
\name{test_fun}
\alias{test_fun}
\title{A Test Function}
\usage{
test_fun(num1, num2)
}
\arguments{
\item{num1}{Any random number.}

\item{num2}{Any random number.}
}
\description{
This function is so dummy.
}
\examples{
test_fun(28,2)
}
\keyword{Num}
